Extract ExpandDims Squeeze reshape pass
